+++ draft-packages/avatar/docs/Avatar.stories.tsx
@@ -37,20 +37,20 @@ export const PhotoPersonal = () => (
   <>
     <Avatar
       size="large"
-      avatarSrc="https://www.fillmurray.com/200/300"
-      fullName="John Smith"
+      avatarSrc="https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png"
+      fullName="Jane Doe"
     />
     <br />
     <Avatar
       size="medium"
-      avatarSrc="https://www.fillmurray.com/200/300"
-      fullName="John Smith"
+      avatarSrc="https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png"
+      fullName="Jane Doe"
     />
     <br />
     <Avatar
       size="small"
-      avatarSrc="https://www.fillmurray.com/200/300"
-      fullName="John Smith"
+      avatarSrc="https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png"
+      fullName="Jane Doe"
     />
   </>
 )