+++ draft-packages/form/docs/ElmSelectField.stories.tsx
@@ -1,10 +1,13 @@
-import { loadElmStories } from "@cultureamp/elm-storybook"
+import { loadElmStories } from "elm-storybook"
 import Ports from "@kaizen/draft-select/KaizenDraft/Select/ports"
+
+const compiledElm = require("../ElmStories/SelectFieldStories.elm").Elm
+  .ElmStories.SelectFieldStories
 
 loadElmStories(
   "SelectField (Elm)",
   module,
-  require("./SelectFieldStories.elm"),
+  compiledElm,
   ["Single", "Multi-Select"],
   Ports
 )