+++ draft-packages/form/docs/ElmToggleSwitchField.stories.tsx
@@ -1,14 +1,12 @@
-import { loadElmStories } from "@cultureamp/elm-storybook"
+import { loadElmStories } from "elm-storybook"
 
-loadElmStories(
-  "ToggleSwitchField (Elm)",
-  module,
-  require("./ToggleSwitchField.stories.elm"),
-  [
-    "Default theme",
-    "Freemium theme",
-    "Disabled Off",
-    "Disabled On",
-    "Label at end",
-  ]
-)
+const compiledElm = require("../ElmStories/ToggleSwitchFieldStories.elm").Elm
+  .ElmStories.ToggleSwitchFieldStories
+
+loadElmStories("ToggleSwitchField (Elm)", module, compiledElm, [
+  "Default theme",
+  "Freemium theme",
+  "Disabled Off",
+  "Disabled On",
+  "Label at end",
+])