+++ draft-packages/popover/docs/Popover.stories.tsx
@@ -1,18 +1,20 @@
-import { DismissiblePositiveAutohide } from "@kaizen/component-library/stories/InlineNotification.stories"
-import { Popover } from "@kaizen/draft-popover"
+import {
+  PopoverLegacy,
+  usePopover,
+  Popover as PopoverRaw,
+} from "@kaizen/draft-popover"
 import * as React from "react"
 import guidanceIcon from "@kaizen/component-library/icons/guidance.icon.svg"
-import { Avatar } from "@kaizen/draft-avatar"
 import { withDesign } from "storybook-addon-designs"
 import { figmaEmbed } from "../../../storybook/helpers"
 
 export default {
   title: "Popover (React)",
-  component: Popover,
+  component: PopoverRaw,
   parameters: {
     info: {
       text: `
-      import { Popover } from "@kaizen/draft-popover"
+      import { usePopover } from "@kaizen/draft-popover"
       `,
     },
     ...figmaEmbed(
@@ -26,212 +28,276 @@ const Container = ({ children }: { children: React.ReactNode }) => (
   <div style={{ minHeight: "150px" }}>{children}</div>
 )
 
-export const DefaultKaizenSiteDemo = () => (
-  <Container>
-    <Popover heading="Default">
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
+const TargetElement = ({
+  referenceElementRef,
+}: {
+  referenceElementRef: (element: HTMLElement | null) => void
+}) => (
+  <div style={{ textAlign: "center" }}>
+    <div
+      ref={referenceElementRef}
+      style={{
+        display: "inline-block",
+        background: "#888",
+        padding: "8px",
+      }}
+    >
+      Target element
+    </div>
+  </div>
 )
+
+export const DefaultKaizenSiteDemo = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover heading="Default">
+        Popover body that explains something useful, is optional, and not
+        critical to completing a task.
+      </Popover>
+    </Container>
+  )
+}
 
 DefaultKaizenSiteDemo.storyName = "Default (Kaizen Site Demo)"
 
-export const DefaultWithoutHeading = () => (
-  <Container>
-    <Popover>
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
+export const DefaultWithoutHeading = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover>
+        Popover body that explains something useful, is optional, and not
+        critical to completing a task.
+      </Popover>
+    </Container>
+  )
+}
 
 DefaultWithoutHeading.storyName = "Default without heading"
 
-export const Informative = () => (
-  <Container>
-    <Popover heading="Informative" variant="informative">
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
+export const Informative = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover heading="Informative" variant="informative">
+        Popover body that explains something useful, is optional, and not
+        critical to completing a task.
+      </Popover>
+    </Container>
+  )
+}
 
-export const InformativeWithSingleLine = () => (
-  <Container>
-    <Popover
-      heading="Informative-default-with-single-line"
-      variant="informative"
-      singleLine
-    >
-      {"http://employee-data.integrations.eu.cultureamp.com/iamaverylongurl/" +
-        "iamaverylongurl/iamaverylongurl/iamaverylongurl"}
-    </Popover>
-  </Container>
-)
+export const InformativeWithSingleLine = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover
+        heading="Informative-default-with-single-line"
+        variant="informative"
+        singleLine
+      >
+        {"http://employee-data.integrations.eu.cultureamp.com/iamaverylongurl/" +
+          "iamaverylongurl/iamaverylongurl/iamaverylongurl"}
+      </Popover>
+    </Container>
+  )
+}
 
 InformativeWithSingleLine.storyName = "Informative with singleLine"
 
-export const InformativeLarge = () => (
-  <Container>
-    <Popover
-      heading="Informative-large-with-multi-line"
-      variant="informative"
-      size="large"
-    >
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
+export const InformativeLarge = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover
+        heading="Informative-large-with-multi-line"
+        variant="informative"
+        size="large"
+      >
+        Popover body that explains something useful, is optional, and not
+        critical to completing a task.
+      </Popover>
+    </Container>
+  )
+}
 
-export const InformativeLargeWithSingleLine = () => (
-  <Container>
-    <Popover
-      heading="Informative-large-with-single-line"
-      variant="informative"
-      size="large"
-      singleLine
-    >
-      {"http://employee-data.integrations.eu.cultureamp.com/iamaverylongurl/" +
-        "iamaverylongurl/iamaverylongurl/iamaverylongurl"}
-    </Popover>
-  </Container>
-)
+export const InformativeLargeWithSingleLine = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover
+        heading="Informative-large-with-single-line"
+        variant="informative"
+        size="large"
+        singleLine
+      >
+        {"http://employee-data.integrations.eu.cultureamp.com/iamaverylongurl/" +
+          "iamaverylongurl/iamaverylongurl/iamaverylongurl"}
+      </Popover>
+    </Container>
+  )
+}
 
 InformativeLargeWithSingleLine.storyName = "Informative Large with singleLine"
 
-export const InformativeWithCustomIcon = () => (
-  <Container>
-    <Popover
-      heading="Informative"
-      variant="informative"
-      customIcon={guidanceIcon}
-    >
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
+export const InformativeWithCustomIcon = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover
+        heading="Informative"
+        variant="informative"
+        customIcon={guidanceIcon}
+      >
+        Popover body that explains something useful, is optional, and not
+        critical to completing a task.
+      </Popover>
+    </Container>
+  )
+}
 
 InformativeWithCustomIcon.storyName = "Informative with a custom icon"
 
-export const Positive = () => (
-  <Container>
-    <Popover heading="Positive" variant="positive">
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
-
-export const Negative = () => (
-  <Container>
-    <Popover heading="Negative" variant="negative">
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
-
-export const Cautionary = () => (
-  <Container>
-    <Popover heading="Cautionary" variant="cautionary">
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
-
-export const Dismissible = () => (
-  <Container>
-    <Popover heading="Dismissible" dismissible>
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-  </Container>
-)
-
-export const ArrowAbove = () => (
-  <Container>
-    <div style={{ marginTop: "1.5rem" }}>
-      <Popover heading="Arrow above" side="top">
+export const Positive = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover heading="Positive" variant="positive">
         Popover body that explains something useful, is optional, and not
         critical to completing a task.
       </Popover>
-    </div>
-  </Container>
-)
+    </Container>
+  )
+}
 
-ArrowAbove.storyName = "Arrow above"
-
-export const ArrowStart = () => (
-  <Container>
-    <div style={{ marginTop: "1.5rem" }}>
-      <Popover heading="Arrow start" position="start">
+export const Negative = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover heading="Negative" variant="negative">
         Popover body that explains something useful, is optional, and not
         critical to completing a task.
       </Popover>
-    </div>
-  </Container>
-)
+    </Container>
+  )
+}
 
-ArrowStart.storyName = "Arrow start"
-
-export const ArrowEnd = () => (
-  <Container>
-    <div style={{ marginTop: "1.5rem" }}>
-      <Popover heading="Arrow end" position="end" side="top">
+export const Cautionary = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover heading="Cautionary" variant="cautionary">
         Popover body that explains something useful, is optional, and not
         critical to completing a task.
       </Popover>
-    </div>
-  </Container>
-)
+    </Container>
+  )
+}
 
-ArrowEnd.storyName = "Arrow end"
+export const Dismissible = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <TargetElement referenceElementRef={referenceElementRef} />
+      <Popover heading="Dismissible" dismissible>
+        Popover body that explains something useful, is optional, and not
+        critical to completing a task.
+      </Popover>
+    </Container>
+  )
+}
 
-export const BoxOffset = () => (
-  <Container>
-    <>
-      <div style={{ marginTop: "1.5rem", height: 200 }}>
-        <Popover
-          heading="Box offset"
-          position="center"
-          side="top"
-          boxOffset={-50}
-        >
+export const PlacementTop = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <div style={{ marginTop: "200px" }}>
+        <TargetElement referenceElementRef={referenceElementRef} />
+        <Popover heading="Placement top" placement="top">
           Popover body that explains something useful, is optional, and not
           critical to completing a task.
         </Popover>
       </div>
-      <div style={{ marginTop: "1.5rem", height: 200 }}>
-        <Popover
-          heading="Box offset"
-          position="center"
-          side="bottom"
-          boxOffset={50}
-        >
+    </Container>
+  )
+}
+
+PlacementTop.storyName = "Placement top"
+
+export const PlacementStart = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <div style={{ marginTop: "1.5rem" }}>
+        <TargetElement referenceElementRef={referenceElementRef} />
+        <Popover heading="Placement start" placement="bottom-start">
           Popover body that explains something useful, is optional, and not
           critical to completing a task.
         </Popover>
       </div>
-    </>
-  </Container>
-)
+    </Container>
+  )
+}
 
-BoxOffset.storyName = "Box offset"
+PlacementStart.storyName = "Placement start"
 
-export const BoxWithYOffset = () => (
+export const PlacementEnd = () => {
+  const [referenceElementRef, Popover] = usePopover()
+  return (
+    <Container>
+      <div style={{ marginTop: "1.5rem" }}>
+        <TargetElement referenceElementRef={referenceElementRef} />
+        <Popover heading="Placement end" placement="bottom-end">
+          Popover body that explains something useful, is optional, and not
+          critical to completing a task.
+        </Popover>
+      </div>
+    </Container>
+  )
+}
+
+PlacementEnd.storyName = "Placement end"
+
+export const LegacyPopover = () => (
   <Container>
-    <Popover heading="Default" boxOffset={{ yOffset: "calc(-100% + -12px)" }}>
-      Popover body that explains something useful, is optional, and not critical
-      to completing a task.
-    </Popover>
-    <div style={{ margin: "200px auto auto", width: "fit-content" }}>
-      <Avatar fullName="Test user"></Avatar>
+    <div
+      style={{
+        position: "relative",
+        height: "200px",
+      }}
+    >
+      <PopoverLegacy heading="Arrow top end" position="end" side="top">
+        The legacy popover gets used when the referenceElement prop is not
+        included.
+      </PopoverLegacy>
+    </div>
+
+    <div
+      style={{
+        position: "relative",
+        height: "200px",
+      }}
+    >
+      <PopoverLegacy
+        heading="Arrow bottom start"
+        position="start"
+        side="bottom"
+      >
+        The legacy popover gets used when the referenceElement prop is not
+        included.
+      </PopoverLegacy>
     </div>
   </Container>
 )
 
-BoxWithYOffset.storyName = "Popover with Y offset positioning"
+LegacyPopover.storyName = "Legacy Popover"