+++ packages/component-library/stories/ElmButton.stories.tsx
@@ -1,16 +1,14 @@
-import { loadElmStories } from "@cultureamp/elm-storybook"
+import { loadElmStories } from "elm-storybook"
 
-loadElmStories(
-  "Button (deprecated) (Elm)",
-  module,
-  require("./Button.stories.elm"),
-  [
-    "Default",
-    "Primary",
-    "Secondary",
-    "Destructive",
-    "Destructive w/ Icon",
-    "Secondary Destructive",
-    "Secondary Destructive w/ Icon",
-  ]
-)
+const compiledElm = require("../ElmStories/ButtonStories.elm").Elm.ElmStories
+  .ButtonStories
+
+loadElmStories("Button (deprecated) (Elm)", module, compiledElm, [
+  "Default",
+  "Primary",
+  "Secondary",
+  "Destructive",
+  "Destructive w/ Icon",
+  "Secondary Destructive",
+  "Secondary Destructive w/ Icon",
+])