+++ draft-packages/badge/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.3.0...@kaizen/draft-badge@1.3.1) (2021-03-15)
+
+
+### Bug Fixes
+
+* add deprecation warnings to ultra legacy style, type and layout mixins + remove internal usage ([#1046](https://github.com/cultureamp/kaizen-design-system/issues/1046)) ([893ba13](https://github.com/cultureamp/kaizen-design-system/commit/893ba134d49468dc1cda3ffd847a056cf4886071))
+
+
+
+
+
+# [1.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.2.5...@kaizen/draft-badge@1.3.0) (2021-03-04)
+
+
+### Features
+
+* Update Badge component to heart theme ([#1173](https://github.com/cultureamp/kaizen-design-system/issues/1173)) ([0aec6f2](https://github.com/cultureamp/kaizen-design-system/commit/0aec6f2f941d5f305a3ae451f6e977e5710a7c49))
+
+
+
+
+
 ## [1.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-badge@1.2.4...@kaizen/draft-badge@1.2.5) (2021-02-25)