+++ draft-packages/card/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@1.2.7...@kaizen/draft-card@1.2.8) (2021-03-15)
+
+
+### Bug Fixes
+
+* upgrade to Elm 0.19.1 ([#1038](https://github.com/cultureamp/kaizen-design-system/issues/1038)) ([07cd9e4](https://github.com/cultureamp/kaizen-design-system/commit/07cd9e4039d5cacfc64f752e1d3a966507ebc377))
+
+
+
+
+
 ## [1.2.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@1.2.6...@kaizen/draft-card@1.2.7) (2021-02-25)