+++ draft-packages/tag/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.8.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.7.18...@kaizen/draft-tag@1.8.0) (2021-03-22)
+
+
+### Features
+
+* Update Tag to support Heart ([#1238](https://github.com/cultureamp/kaizen-design-system/issues/1238)) ([d03ca5e](https://github.com/cultureamp/kaizen-design-system/commit/d03ca5e1298a2b8dcc6185d2c4c17156fb19fa5f))
+
+
+
+
+
+## [1.7.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.7.17...@kaizen/draft-tag@1.7.18) (2021-03-17)
+
+**Note:** Version bump only for package @kaizen/draft-tag
+
+
+
+
+
+## [1.7.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.7.16...@kaizen/draft-tag@1.7.17) (2021-03-15)
+
+
+### Bug Fixes
+
+* add deprecation warnings to ultra legacy style, type and layout mixins + remove internal usage ([#1046](https://github.com/cultureamp/kaizen-design-system/issues/1046)) ([893ba13](https://github.com/cultureamp/kaizen-design-system/commit/893ba134d49468dc1cda3ffd847a056cf4886071))
+
+
+
+
+
+## [1.7.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.7.15...@kaizen/draft-tag@1.7.16) (2021-03-15)
+
+
+### Bug Fixes
+
+* upgrade to Elm 0.19.1 ([#1038](https://github.com/cultureamp/kaizen-design-system/issues/1038)) ([07cd9e4](https://github.com/cultureamp/kaizen-design-system/commit/07cd9e4039d5cacfc64f752e1d3a966507ebc377))
+
+
+
+
+
 ## [1.7.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@1.7.14...@kaizen/draft-tag@1.7.15) (2021-02-26)
 
 **Note:** Version bump only for package @kaizen/draft-tag